refactor(spaces): the settings pop-up is its own component, so a second host is possible
No behaviour change. Groundwork for U-1: a cog at the far right of the Brain tab
strip, opening these settings for the space you are already looking at.

The modal was 84 lines of template inside `spaces.component`, reachable only from
that page's own table. It is driven by ONE signal — set
`SpaceSettingsState.settingsSpace(space)` and it appears — so hosting it elsewhere
was only ever a question of who provides the service.

## The template was moved, not retyped

By asserted line range, then compared byte-for-byte against `HEAD` and confirmed
present verbatim in the new component. A refactor that retypes eighty lines of
template is a refactor that loses one of them, and a modal that renders subtly
wrong passes every build.

## What came along, and what deliberately did not

`governedBy`, `saveSettings` and `attemptClose` exist only to serve this template,
so they moved.

`canLeave` and the `beforeunload` handler stayed on the Spaces page: they are
ROUTE concerns, and a modal that can be opened from two pages must not own either
page's navigation guard.

The discard prompt moved to the SERVICE rather than being copied, because both the
(X) and the route guard need the identical question. It also folds in the
`isDirty()` check, so no caller can pair it with its own and invert one.

## Two corrections to what I said before starting

There is no second CSS location. All seven `.sp-*` rules are in
`space-dialog.styles.ts` and nowhere else; my claim that five were inline came
from a grep matching the `<!-- sp-body -->` closing COMMENTS in the template.

And the first attempt at this failed. It precomputed four absolute line ranges for
the class members and applied them in sequence, so each cut shifted the ones after
it. `tsc` was silent; Angular reported every import as unused, which is what a
broken template string looks like from outside. That branch was discarded and main
was never touched. Every removal here is anchored on the member's own opening line
and the closing brace at its indentation, re-derived between each.

Also walked into my own documented trap on the way: the placeholder comment left
behind in the host template contained backticks, which ends the template string.

## Tests followed the code

Two describes moved to `space-settings-popup.component.spec.ts` — they were always
testing the modal rather than the page around it. Their setup differs in one way
that matters and previews the Brain host: the pop-up does not PROVIDE
`SpaceSettingsState` or `SpacesStore`, so the spec provides them.

One test went with neither: it drove the page's create dialog, which is the page's
business.

`space-purpose-one-field` pinned the vote-pending branch by path and has been
re-pointed. Its first assertion is positive, so a path resolving to a file without
that branch fails loudly rather than passing vacuously.
