+++ client/src/app/pages/settings/spaces.component.ts
@@ -21,7 +21,7 @@ import { SummaryStripComponent, type SummaryItem } from '../../shared/summary-st
 import { SpaceSettingsState, type TypeSchemaState } from './space-settings-state.service';
 import { SpacesStore } from './spaces-store.service';
 import { SPACE_DIALOG_STYLES } from './space-dialog.styles';
-import { SpaceSettingsTabComponent } from './space-settings-tab.component';
+import { SpaceSettingsPopupComponent } from './space-settings-popup.component';
 import { SpaceDuplicatesTabComponent } from './space-duplicates-tab.component';
 import { SpaceDangerTabComponent } from './space-danger-tab.component';
 import { SpaceSchemaTabComponent } from './space-schema-tab.component';
@@ -34,7 +34,7 @@ import { StatusPillComponent } from '../../shared/status-pill.component';
   selector: 'app-spaces',
   standalone: true,
   imports: [ProxySpaceBadgeComponent, CommonModule, FormsModule, TranslocoPipe, DragDropModule, PhIconComponent, SummaryStripComponent,
-    SpaceSettingsTabComponent, SpaceDuplicatesTabComponent, SpaceDangerTabComponent, SpaceSchemaTabComponent,
+    SpaceSettingsPopupComponent,
     SpaceCreateDialogComponent, ModalDirective, HscrollTopDirective, StatusPillComponent],
   // Provided here (not root) so each mount gets its own settings state, with a lifetime tied to
   // this component rather than the app.
@@ -47,90 +47,8 @@ import { StatusPillComponent } from '../../shared/status-pill.component';
       <app-space-create-dialog (closed)="showCreateDialog.set(false)" />
     }
 
-    <!-- SETTINGS POPUP -->
-    @if (state.settingsSpace()) {
-      <div class="sp-backdrop">
-        <div class="sp-panel" [appModal]="state.settingsSpace()!.label" (dismiss)="attemptClose()">
-          <div class="sp-header">
-            <div style="flex:1;min-width:0;">
-              <div style="font-weight:600;font-size:16px;overflow:hidden;text-overflow:ellipsis;white-space:nowrap;">{{ state.settingsSpace()!.label }}</div>
-              <div style="font-size:12px;color:var(--text-muted);font-family:var(--font-mono);">{{ state.settingsSpace()!.id }}</div>
-            </div>
-            <!-- Governed BEFORE you type, not after you press Save.
-                 Saving a networked space answers 202 vote_pending: the change is submitted for a vote
-                 rather than applied. That used to be discovered by pressing the button — the notice
-                 explains it afterwards, which is the wrong end of the interaction to learn it. The
-                 membership is already on the space record (its networks array), so this costs no request. -->
-            @if (governedBy(); as nets) {
-              <!-- The link icon, because that is what the sidebar already uses for Networks; the pill has
-                   to read as the same concept, not a new one. (No users/gavel icon is registered, and an
-                   unregistered name renders BLANK with no error.) -->
-              <app-status-pill variant="pending" icon="link"
-                [attr.title]="'spaces.popup.governedHint' | transloco: { networks: nets }">
-                {{ 'spaces.popup.governed' | transloco }}
-              </app-status-pill>
-            }
-            <button class="icon-btn" [attr.aria-label]="'common.close' | transloco" (click)="attemptClose()"><ph-icon name="x" [size]="14"/></button>
-          </div>
-          <div class="sp-tabs" role="tablist" [attr.aria-label]="'spaces.settings.tabsAriaLabel' | transloco">
-            <button class="sp-tab" [class.active]="state.settingsTab()==='settings'" [attr.aria-selected]="state.settingsTab()==='settings'" role="tab" (click)="state.settingsTab.set('settings')">{{ 'spaces.popup.tab.settings' | transloco }}</button>
-            <button class="sp-tab" [class.active]="state.settingsTab()==='schema'" [attr.aria-selected]="state.settingsTab()==='schema'" role="tab"   (click)="state.settingsTab.set('schema')">{{ 'spaces.popup.tab.schema' | transloco }}</button>
-            <button class="sp-tab" [class.active]="state.settingsTab()==='duplicates'" [attr.aria-selected]="state.settingsTab()==='duplicates'" role="tab" (click)="state.settingsTab.set('duplicates')">{{ 'spaces.popup.tab.duplicates' | transloco }}</button>
-            <button class="sp-tab danger-tab" [class.active]="state.settingsTab()==='danger'" [attr.aria-selected]="state.settingsTab()==='danger'" role="tab" (click)="state.settingsTab.set('danger')">{{ 'spaces.popup.tab.dangerZone' | transloco }}</button>
-          </div>
-          <div class="sp-body">
-
-            <!-- SETTINGS TAB -->
-            @if (state.settingsTab() === 'settings') {
-              <app-space-settings-tab />
-            }
-
-            <!-- SCHEMA TAB -->
-            @if (state.settingsTab() === 'schema') {
-              <app-space-schema-tab />
-            }
-
-            <!-- DUPLICATES TAB -->
-            @if (state.settingsTab() === 'duplicates') {
-              <app-space-duplicates-tab />
-            }
-
-            <!-- DANGER ZONE TAB -->
-            @if (state.settingsTab() === 'danger') {
-              <app-space-danger-tab />
-            }
-          </div><!-- sp-body -->
-
-          @if (state.settingsTab() !== 'danger' && state.settingsTab() !== 'duplicates') {
-            <div class="sp-footer">
-              @if (state.settingsError()) {
-                <div class="alert alert-error" style="flex:1;margin:0;padding:6px 12px;font-size:13px;">{{ state.settingsError() }}</div>
-              }
-              @if (state.settingsNotice()) {
-                <div class="alert alert-info" style="flex:1;margin:0;padding:6px 12px;font-size:13px;">{{ state.settingsNotice() }}</div>
-              }
-              <!-- Once the outcome is TERMINAL, the button says so.
-                   A governed save answers 202 and opens a vote, so the work is finished and there is nothing
-                   left to submit, but the button still read "Save changes" and the only exit was the (X),
-                   which universally means DISCARD. A reporting operator: "i have to click (X) which feels
-                   unsure if the changes are now actually up for vote or discarded."
-                   That is a wrong-action risk, not a wobble: read as cancel, someone looks for another way to
-                   confirm, saves again, and creates a SECOND proposal for the same change. A button that
-                   submitted successfully must not still be offering to submit. -->
-              @if (state.settingsNotice()) {
-                <button class="btn btn-primary" type="button" (click)="state.closeSettings()">
-                  {{ 'spaces.popup.footer.done' | transloco }}
-                </button>
-              } @else {
-                <button class="btn btn-primary" type="button" (click)="saveSettings()" [disabled]="state.settingsSaving()">
-                  @if (state.settingsSaving()) { <span class="spinner" style="width:12px;height:12px;border-width:2px;"></span> }{{ 'spaces.popup.footer.saveChanges' | transloco }}
-                </button>
-              }
-            </div>
-          }
-        </div><!-- sp-panel -->
-      </div><!-- sp-backdrop -->
-    }
+    <!-- The space settings pop-up: its own component, opened by setting state.settingsSpace(). -->
+    <app-space-settings-popup />
 
     <!-- Import conflict dialog -->
 
@@ -278,21 +196,6 @@ export class SpacesComponent implements OnInit {
   });
 
   /**
-   * The networks governing the open space, as a readable list — or null when it is governed by none.
-   *
-   * Read from the space record the list endpoint already returns (`networks`), so the badge costs no
-   * request and cannot disagree with the chip the Brain shows for the same space.
-   *
-   * Deliberately not keyed on `networkStatus`: that reports whether something is *happening* (a vote, a
-   * sync, a degraded peer), and a quiet network still means Save opens a round. The question the badge
-   * answers is "is this space governed", which is membership.
-   */
-  governedBy = computed(() => {
-    const nets = this.state.settingsSpace()?.networks ?? [];
-    return nets.length > 0 ? nets.map(n => n.label).join(', ') : null;
-  });
-
-  /**
    * Usage per space id, over the last 7 days, from ONE request.
    *
    * A week, not a day: usefulness is a question about a habit, and a space queried every Monday reads as dead
@@ -382,77 +285,14 @@ export class SpacesComponent implements OnInit {
     return `${gib.toFixed(2)} GiB`;
   }
 
-  saveSettings(): void {
-    const target = this.state.settingsSpace();
-    if (!target) return;
-    this.state.settingsSaving.set(true);
-    this.state.settingsError.set('');
-    this.state.settingsNotice.set('');
-    this.spacesApi.updateSpace(target.id, {
-      label:  this.state.stForm.label.trim() || target.label,
-      maxGiB: this.state.stForm.maxGiB,
-      // NO recordTtlDays. It moved to the Danger Zone, which saves itself; this tab has only a note pointing
-      // there. Echoing the stored value back was harmless while it was one number and is not now: the space
-      // tier is five buckets, and a scalar write REPLACES the whole object — so a label edit would have
-      // flattened every per-collection window to one figure.
-      documentExtraction: this.state.stForm.documentExtraction || null, // F11-c ('' = inherit instance default)
-      imageAnalysis: this.state.stForm.imageAnalysis || null,           // '' = inherit instance default
-      audioAnalysis: this.state.stForm.audioAnalysis || null,
-      videoAnalysis: this.state.stForm.videoAnalysis || null,
-      textAnalysis: this.state.stForm.textAnalysis || null,
-      meta:   this.state.buildMeta(),
-      // Save persists the state the editor is showing. Without this the PATCH merges, so a type deleted in
-      // the UI is simply not mentioned and the server keeps it — the delete appears to work, survives the
-      // save, and is still there on reload.
-      typeSchemasMode: 'replace',
-    }).subscribe({
-      next: (result) => {
-        this.state.settingsSaving.set(false);
-        // A networked space does not apply a meta change on the spot: the server opens a vote round per
-        // network and answers 202 with no `space`. Destructuring it as one threw inside this callback —
-        // which RxJS does not route to `error` — so Save appeared to do nothing at all, and the editor
-        // then asked whether to discard the change it had just submitted. Say what happened instead.
-        if (result.status === 'vote_pending') {
-          this.state.settingsNotice.set(this.transloco.translate('spaces.settings.votePending', {
-            networks: result.rounds.map(r => r.networkLabel).join(', '),
-          }));
-          this.state.markPristine();   // it is submitted; it is not an unsaved edit any more
-          return;                      // stay open so the notice is read, unlike the applied path
-        }
-        this.store.applySpace(result.space);
-        // Re-baseline BEFORE closing. The dirty snapshot was only ever taken when a space was opened, so
-        // a successful save left it stale: the editor still compared against the pre-save values and
-        // reported unsaved changes for edits that were already persisted. Closing here happened to hide
-        // it, but any path that keeps the dialog open (or reopens it without a full load) nagged — and a
-        // discard prompt after a save teaches people to click through discard prompts.
-        this.state.markPristine();
-        this.state.closeSettings();
-      },
-      error: (err) => { this.state.settingsSaving.set(false); this.state.settingsError.set(err.error?.error ?? this.transloco.translate('spaces.error.saveFailed')); },
-    });
-  }
-
   // ── Unsaved-changes guard (U4) ─────────────────────────────────────────────
-
-  /** Close the settings dialog, prompting first if the editor has unsaved edits. */
-  async attemptClose(): Promise<void> {
-    if (this.state.isDirty() && !(await this.confirmDiscard())) return;
-    this.state.closeSettings();
-  }
 
   /** CanDeactivate hook: block leaving the Spaces route while the editor has unsaved edits. */
   canLeave(): boolean | Promise<boolean> {
-    return this.state.isDirty() ? this.confirmDiscard() : true;
-  }
-
-  private confirmDiscard(): Promise<boolean> {
-    return this.confirmDialog.confirm({
-      title:   this.transloco.translate('spaces.unsaved.title'),
-      message: this.transloco.translate('spaces.unsaved.message'),
-      confirmLabel: this.transloco.translate('spaces.unsaved.confirm'),
-      cancelLabel:  this.transloco.translate('spaces.unsaved.cancel'),
-      danger: true,
-    });
+    // The prompt lives on the service now: the pop-up's (X) needs the identical question, and two
+    // copies of "are you sure you want to lose these edits" is two places for the answer to drift.
+    // This hook stays here — a modal openable from two pages must not own either page's guard.
+    return this.state.confirmDiscardIfDirty();
   }
 
   /** Native prompt on reload/tab-close while dirty — EventSource-style dialogs aren't allowed here. */
